Correction: E2 partitioner key type correct to Int to match new row and col ID type.

+++ src/main/scala/edu/asu/sparkpartitioning/utils/MatrixPartitioners.scala
@@ -13,7 +13,7 @@ object MatrixPartitioners {
     override def numPartitions: Int = numberOfParts
 
     override def getPartition(key: Any): Int =
-      (key.asInstanceOf[Long] % numberOfParts.toLong).toInt
+      key.asInstanceOf[Int] % numberOfParts
   }
 
 }
